[FIX] Moved OpenDatabase
